Update package weight unit and add activationId field
Updated the package weight unit from integer to floating number in the PackageItem class. Also, added an activationId field in the CourierServiceDTO class. These modifications ensure better precision for package weight and enhance order tracking.

+++ src/Types/PackageItem.php
@@ -5,10 +5,10 @@ namespace Ag84ark\ColeteOnlineRoPhp\Types;
 class PackageItem extends BaseType
 {
     public function __construct(
-        protected int $weight,
-        protected ?int $height,
-        protected ?int $width,
-        protected ?int $length,
+        protected float $weight,
+        protected ?float $height,
+        protected ?float $width,
+        protected ?float $length,
     ) {
     }
